requirements.txt
pip==22.3.1
click==8.1.3
Flask==2.2.2
gunicorn==20.1.0
itsdangerous==2.1.2
jinja2==3.1.2
markupsafe==2.1.2
werkzeug==2.2.2
pillow==9.3.0